chore: move database and visualisation to output submodule
