+++ rex_xai/rex.py
@@ -16,7 +16,7 @@ from sqlalchemy.orm import Session
 from tqdm import trange  # type: ignore
 
 from rex_xai.input.config import CausalArgs
-from rex_xai.database import update_database
+from rex_xai.output.database import update_database
 from rex_xai.evaluation import Evaluation
 from rex_xai.multi_explanation import MultiExplanation
 from rex_xai.explanation import Explanation
